:sparkles: implemented float16

+++ include/a_funcs.hpp
@@ -1,22 +1,24 @@
 #pragma once
+#include "../third_party/half/half.hpp"
+
 constexpr auto A = 0.0001;
 
 namespace rum
 {
     //relu
 
-    float Relu(float x)
+    FLOAT16 Relu(FLOAT16 x)
     {
         return (x > 0) * x;
     }
-    float ReluPrime(float x)
+    FLOAT16 ReluPrime(FLOAT16 x)
     {
         return x > 0;
     }
 
     //leaky relu
 
-    float ReluLeaky(float x)
+    FLOAT16 ReluLeaky(FLOAT16 x)
     {
         if (x < 0)
         {
@@ -28,7 +30,7 @@ namespace rum
         }
     }
 
-    float ReluLeakyPrime(float x)
+    FLOAT16 ReluLeakyPrime(FLOAT16 x)
     {
         if (x < 0)
         {
@@ -42,35 +44,35 @@ namespace rum
 
     //tanh
 
-    float Tanh(float x)
+    FLOAT16 Tanh(FLOAT16 x)
     {
         return tanh(x);
     }
-    float TanhPrime(float x)
+    FLOAT16 TanhPrime(FLOAT16 x)
     {
         return sinh(x) / cosh(x);
     }
 
     //sigmoid
 
-    float Sigmoid(float x)
+    FLOAT16 Sigmoid(FLOAT16 x)
     {
         return 1 / (1 + exp(-x));
     }
 
-    float SigmoidPrime(float x)
+    FLOAT16 SigmoidPrime(FLOAT16 x)
     {
         return exp(-x) / std::pow(1 + exp(-x), 2);
     }
 
     //swish
 
-    float Swish(float x)
+    FLOAT16 Swish(FLOAT16 x)
     {
         return x * Sigmoid(x);
     }
 
-    float SwishPrime(float x)
+    FLOAT16 SwishPrime(FLOAT16 x)
     {
         return (exp(x) * (exp(x) + x + 1)) / std::pow(exp(x) + 1, 2);
     }
